move files around, add replacer state
