+++ src/lib/state/replacer.svelte.ts
@@ -1,0 +1,4 @@
+import defaultReplacer from "./defaultReplacer";
+import type { Replacer } from "./replacer.d";
+
+const replacer = $state(defaultReplacer);
